requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
rich 
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib
alembic
pydantic_settings
pydantic[email]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
redis>=4.5.0
langchain_openai

# For documentation
markdown2